% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lergm.R
\name{lergm}
\alias{lergm}
\title{Estimation of ERGMs using Exact likelihood functions}
\usage{
lergm(formula, control = list(maxit = 1000, reltol = 1e-30),
  allstats_force = TRUE, offset = NULL, stats = NULL)
}
\arguments{
\item{formula, control, offset}{See \link[ergm:ergm]{ergm::ergm}.}

\item{allstats_force}{Logical. Passed to \link[ergm:ergm.allstats]{ergm::ergm.allstats}.}

\item{astats}{List as returned by \link[ergm:ergm.allstats]{ergm::ergm.allstats}. When this is provided,
the function does not call \code{ergm.allstats}, which can be useful in simulations.}
}
\value{
An object of class \code{ergm} and \code{lergm}.
}
\description{
As a difference from \link[=ergm]{ergm::ergm}, \code{lergm} uses the exact log-likelihood
function for fitting the model. This implies that all the \code{2^(n*(n-1))}
graphs are generated for computing the normalizing constant of the ERGM
model. This implies that models with up to 5 nodes are relatively simple
to fit, since more than that can become infeasible.
}
\examples{

# Generating a small graph
set.seed(12)
n <- 4
net <- sna::rgraph(n, tprob = .7)

model <- net ~ edges + mutual + balance

library(ergm)
ans_lergm <- lergm(model)
ans_ergm  <- ergm(model)

# The lergm should have a larger value
ergm.exact(ans_lergm$coef, model)
ergm.exact(ans_ergm$coef, model)

summary(ans_lergm)
summary(ans_ergm)
}
