% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/similarity.R
\name{similarity}
\alias{similarity}
\alias{Michael}
\alias{Peirce}
\alias{similarity.list}
\alias{similarity.matrix}
\title{Similarity indexes}
\usage{
similarity(M, ..., statistic, normalized = TRUE)

\method{similarity}{list}(M, ..., statistic, normalized = TRUE)

\method{similarity}{matrix}(M, ..., statistic, normalized = TRUE)
}
\arguments{
\item{M}{Either a list of matrices of size `n` (need not to be square), or
a single matrix of size `n` (see details).}

\item{...}{More matrices to be passed to the function.}

\item{statistic}{Character. Name of the similarity index to be using.}

\item{bool}{Logical. When `TRUE` it returns the normalized hamming distance,
which ranges between 0 and 1.}
}
\value{
A matrix of size `n*(n - 1)/2` by 3, where columns 1 and 2 indicate
ego and alter, and the third column the `S14` statistic.
}
\description{
Includes the S14 statistic and hamming distance between two or more
pairs of matrices.
}
\details{
(68) S Michael

(73) Peirce

If `M` is a matrix, then the function requires the user to pass at least a
second matrix via de `...` notation.

In the case of the `S14` function, following Krackhardt's 1989:
\tabular{lcc}{
     \tab R2 \tab   \cr
  R1 \tab a  \tab b \cr
     \tab c  \tab d  
}

\deqn{%
\sqrt{\left(\frac{a}{(a + c)} - \frac{b}{(b + d)}\right)\times\left(\frac{a}{(a + b)} - \frac{c}{(c + d)}\right)}
}{%
S14 = [(a/(a + c) - b/(b + d))*(a/(a + b) - c/(c + d))]^(1/2)
}

Which is an statistic lying between 0 and 1.
}
\examples{
# Getting all networks of size 3
nets <- powerset(3)

# We can compute it over the entire set
head(similarity(nets, statistic="S14"))

# Or over two pairs
similarity(nets[[1]], nets[[2]], nets[[3]], statistic="S14")

# We can do the same with the hamming distance
similarity(nets, statistic="hamming")

}
\references{
Krackhardt, "Assessing the Political Landscape: Structure, Cognition, and Power in Organizations",
Administrative Science Quarterly, Vol. 35, No. 2 (Jun., 1990), pp. 342-369
}
