% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\name{powerset}
\alias{powerset}
\title{Power set of Directed Graphs of size \code{n}}
\usage{
powerset(n, force = FALSE, mc_cores = getOption("mc.cores", 2L),
  chunk_size = 1e+05)
}
\arguments{
\item{n}{Integer. Number of edges.}

\item{force}{Logical. When \code{TRUE} it generates the powerset for \code{n>5}, otherwise
it returns with error.}
}
\description{
Power set of Directed Graphs of size \code{n}
}
\examples{
powerset(2)
}
